feat(postgres): move postgres out to it's own endpoint, BREAKING CHANGE
