+++ package.json
@@ -9,6 +9,10 @@
 			"import": "./src/index.js",
 			"types": "./types/src/index.d.ts"
 		},
+		"./postgres": {
+			"import": "./src/postgresjs",
+			"types": "./types/src/index.d.ts"
+		},
 		"./utils/*": {
 			"import": "./src/utils/*.js",
 			"types": "./types/src/utils/*.d.ts"
@@ -20,13 +24,11 @@
 	},
 	"scripts": {
 		"prepare": "is-ci || husky install",
-		"pretest": "npm run lint",
 		"check-types": "tsc -p ./tsconfig.json",
-		"prepublish": "npm run check-types",
-		"test": "npm run spec && ((c8 report --reporter=text-lcov | coveralls) || exit 0)",
-		"test:ci": "eslint ./ && c8 node --test test/specs/**/*.spec.js && (c8 report --reporter=text-lcov | coveralls)",
+		"test": "npm run lint && npm run spec && ((c8 report --reporter=text-lcov | coveralls) || exit 0)",
+		"test:ci": "npm run lint && c8 node --test test/specs/**/*.spec.js && (c8 report --reporter=text-lcov | coveralls)",
 		"test:integration": "bash ./test/integration/run.sh",
-		"spec": "c8 node --test test/specs/**/*.spec.js",
+		"spec": "c8 node --test 'test/specs/**/*.spec.js'",
 		"lint": "eslint ./ && npx prettier --check . && npm run check-types",
 		"prettier": "prettier --write --ignore-unknown .",
 		"lint-fix": "eslint --fix ./",
