+++ test/specs/filter_reducer.spec.js
@@ -201,77 +201,4 @@ describe('Filter Reducer', () => {
 			});
 		});
 	});
-
-	describe('postgres engine version handling', () => {
-		const engine = 'postgres:16.3';
-
-		it('should search fulltext - with an index', async () => {
-			const dareInst = dareInstance.use({engine});
-
-			const filter = {
-				'*vector_index': 'string',
-			};
-
-			const sql = `a.vector_index @@ to_tsquery('english', ?)`;
-			const values = ['string'];
-
-			const [query] = reduceConditions(filter, {
-				extract,
-				sql_alias: 'a',
-				table_schema,
-				conditional_operators_in_value,
-				dareInstance: dareInst,
-			});
-
-			assert.strictEqual(query.sql, sql);
-			assert.deepStrictEqual(query.values, values);
-		});
-
-		it('should search fulltext - and build an index', async () => {
-			const dareInst = dareInstance.use({engine});
-
-			const filter = {
-				'*givenname,lastname,email': 'string',
-			};
-
-			const sql = `TO_TSVECTOR(a.givenname || ' ' || a.lastname || ' ' || a.email) @@ to_tsquery('english', ?)`;
-			const values = ['string'];
-
-			const [query] = reduceConditions(filter, {
-				extract,
-				sql_alias: 'a',
-				table_schema,
-				conditional_operators_in_value,
-				dareInstance: dareInst,
-			});
-
-			assert.strictEqual(query.sql, sql);
-			assert.deepStrictEqual(query.values, values);
-		});
-
-		 it(`quote json number and boolean values`, () => {
-			const dareInst = dareInstance.use({engine});
-
-			const filter = {
-				jsonSettings: {
-					key: 1,
-					'%str': 'string%',
-				},
-			};
-
-			const [query] = reduceConditions(filter, {
-				extract,
-				sql_alias: 'a',
-				table_schema,
-				conditional_operators_in_value,
-				dareInstance: dareInst,
-			});
-
-			const sql = `(a.jsonSettings->>? = ? AND a.jsonSettings->>? ILIKE ?)`;
-			const values = ['key', '1', 'str', 'string%'];
-
-			assert.strictEqual(query.sql, sql);
-			assert.deepStrictEqual(query.values, values);
-		});
-	});
 });
