+++ test/specs/post.spec.js
@@ -314,46 +314,4 @@ describe('post', () => {
 			});
 		});
 	});
-
-	describe('DB Engine specific tests', () => {
-		const DB_ENGINE = 'postgres:16.3';
-		let dareInst;
-
-		beforeEach(() => {
-			dareInst = dare.use({engine: DB_ENGINE});
-		});
-
-		it(`${DB_ENGINE} should use ON CONFLICT ... UPDATE ...`, async () => {
-			dareInst.execute = async ({sql, values}) => {
-				sqlEqual(
-					sql,
-					'INSERT INTO test ("id", "name") VALUES (?, ?) ON CONFLICT (id) DO UPDATE SET "name"=EXCLUDED."name" RETURNING id'
-				);
-				assert.deepStrictEqual(values, [1, 'name']);
-				return {success: true};
-			};
-
-			return dareInst.post({
-				table: 'test',
-				body: {id: 1, name: 'name'},
-				duplicate_keys_update: ['name'],
-			});
-		});
-		it(`${DB_ENGINE} should use ON CONFLICT DO NOTHING`, async () => {
-			dareInst.execute = async ({sql, values}) => {
-				sqlEqual(
-					sql,
-					'INSERT INTO test ("id", "name") VALUES (?, ?) ON CONFLICT DO NOTHING RETURNING id'
-				);
-				assert.deepStrictEqual(values, [1, 'name']);
-				return {success: true};
-			};
-
-			return dareInst.post({
-				table: 'test',
-				body: {id: 1, name: 'name'},
-				duplicate_keys: 'ignore',
-			});
-		});
-	});
 });
